requirements.txt
python-telegram-bot>=20.0
python-dotenv>=1.0.0
PyYAML>=6.0

deepgram-sdk==4.8.1
